Fixed some issues with setting cooling power and fan speed
The values for cooling power and fan speed was switched up.  I also remoed a check for the local value was the same as the one that is calculated, since that would make it only update if the phone knows it should update. Since the fan can be controlled by a button or reset with a power cycle, we have to assume that it can get out of sync, so sending the bluetooth commands on every interval would make sure it gets back into sync.