Fixed issues with Black Shark phone cooler integration (#125)
* Fixed some issues with setting cooling power and fan speed

The values for cooling power and fan speed was switched up.  I also remoed a check for the local value was the same as the one that is calculated, since that would make it only update if the phone knows it should update. Since the fan can be controlled by a button or reset with a power cycle, we have to assume that it can get out of sync, so sending the bluetooth commands on every interval would make sure it gets back into sync.

* Fixed Brightness and LED issues

- Brightness was set to default as 1 out of a 0-100 range. So set it 100 for full brightness. users can adjust it after.
- Created new extension function toStandardRGB which supplements the color and toRgbp value. It seems like apples "Standard" RGB is actually extended RGB where the value range can go above 1 and belox 0, which breaks stuff. The toStandardRgB clamps this to prevent issues further down with integer conversions. The new function is used for the LED settings. Main reason for a new function was to not break other parts of the code that uses toRgb() which may relie on that extended RGB.