move IPv6 URLs to subpages for SRT(LA) Server, RTMP Server and Moblink Streamer
Most users wont need the IPv6 URLs, and they look very confusing for most people, also the list is very long when it includes the IPv6 addresses. Moved all IPv6 addresses to subpages for advanced users who know what they are doing, and make the UI cleaner this way for most of the normal users.